Move rtg related moudles to rtg/
